requirements.txt
PyQt6>=6.6.0
openai>=1.0.0
pyobjc-framework-Quartz>=10.0
pyobjc-framework-Cocoa>=10.0
psutil>=5.9.0
